requirements.txt
streamlit
pytesseract
pdfplumber
pydantic
pillow
matplotlib
pandas
